+++ CHANGELOG.md
@@ -43,6 +43,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [#31](https://github.com/allora-network/allora-offchain-node/pull/31) SubmitTx fix: if set to false but properly configured, it should still not submit.
+* [#37](https://github.com/allora-network/allora-offchain-node/pull/37) Fix covering nil pointer when params are not available
+* [#38](https://github.com/allora-network/allora-offchain-node/pull/38) Fix error handling (nil pointer dereference) on registration.
+* [#40](https://github.com/allora-network/allora-offchain-node/pull/40) Forecasting fixes
+* [#41](https://github.com/allora-network/allora-offchain-node/pull/41) MSE insteead of MAE, Reputer data validation, refactoring.
+* [#42](https://github.com/allora-network/allora-offchain-node/pull/41) Update to v0.4.0 version of the chain. This contains breaking changes in types.
+
 ## v0.2.0
 
 ### Added